move absolute path requirement for vfs into type system
